requirements.txt
# Django & REST Framework
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# JWT Authentication
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# File Handling
Pillow==10.2.0

# Development & Testing
pytest==8.0.1
pytest-django==4.8.0
coverage==7.4.1

# Documentation
drf-yasg==1.21.7

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Payment Integration
stripe==7.11.0
